requirements.txt
gdown
jiwer
nltk
pyter3
rouge-score
datasets
tensorflow>=1.15
git+https://github.com/google-research/bleurt.git@master

